Rename translator to match style
