+++ web-rdfs/CMakeLists.txt
@@ -22,7 +22,7 @@ target_link_libraries(web_rdfs_server ${Boost_LIBRARIES})
 add_library(http_handlers source/http_handlers.cc)
 target_link_libraries(http_handlers easylogging++)
 
-add_library(web_request_translator source/webRequestTranslator.cc)
+add_library(web_request_translator source/web_request_translator.cc)
 target_link_libraries(web_request_translator easylogging++)
 target_link_libraries(web_request_translator protobuf)
