+++ web-rdfs/include/web_request_translator.h
