+++ web-rdfs/source/web_request_translator.cc
@@ -1,4 +1,4 @@
-#include "webRequestTranslator.h"
+#include "web_request_translator.h"
 
 namespace webRequestTranslator {
     /**
